requirements.txt
# ESP32-WiFi-Sensing-2 - Python Dependencies
# Cài đặt bằng lệnh: pip install -r requirements.txt

# Thu thập dữ liệu từ Serial
pyserial>=3.5

# Xử lý tín hiệu (Stage 2)
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Nhận diện nhịp thở (Stage 3) - đã bao gồm trong scipy

# Kết nối MQTT (Stage 4)
paho-mqtt>=1.6.0

# Phân tích dữ liệu CSV (Stage 1: verify_data.py)
pandas>=1.3.0
